+++ src/shared/copy-static-json.js
@@ -1,10 +1,8 @@
 let cp = require('./copy')
-let fs = require('fs')
-let path = require('path')
+let { existsSync } = require('fs')
+let { join } = require('path')
 let series = require('run-series')
-let getBasePaths = require('./get-base-paths')
 let print = require('../_printer')
-let { readArc } = require('@architect/utils')
 
 /**
  * copies public/static.json
@@ -15,18 +13,14 @@ let { readArc } = require('@architect/utils')
  * nodejs*  | node_modules/@architect/shared/static.json
  * else     | vendor/shared/static.json
  */
-module.exports = function copyStatic (params, callback) {
-  let { update, only } = params
+module.exports = function copyStatic (params, paths, callback) {
+  let { inventory, update, only } = params
+  let { get } = inventory
   let go = !only || only === 'staticJson' || only === 'shared'
 
-  let { arc } = readArc()
-  let staticDir = 'public'
-  if (arc.static && arc.static.some(i => i[0] === 'folder')) {
-    staticDir = arc.static[arc.static.findIndex(i => i[0] === 'folder')][1] || 'public'
-  }
-  let static = path.join(process.cwd(), staticDir, 'static.json')
-  let hasStatic = fs.existsSync(static)
-
+  let staticDir = get.static('folder')
+  let static = staticDir && join(process.cwd(), staticDir, 'static.json')
+  let hasStatic = staticDir && existsSync(static)
   if (hasStatic && go) {
     // Kick off logging
     let done = `Hydrated app with static.json`
@@ -34,23 +28,14 @@ module.exports = function copyStatic (params, callback) {
 
     function _done (err) {
       let cmd = 'copy'
-      if (err) {
-        print({ cmd, err, start, update }, callback)
-      }
-      else {
-        print({ cmd, start, done, update }, callback)
-      }
+      if (err) print({ cmd, err, start, update }, callback)
+      else print({ cmd, start, done, update }, callback)
     }
-    getBasePaths('static', function gotBasePaths (err, paths) {
-      if (err) _done(err)
-      else {
-        series(paths.map(dest => {
-          return function copier (callback) {
-            cp(static, path.join(dest, 'shared', 'static.json'), params, callback)
-          }
-        }), _done)
+    series(paths.map(dest => {
+      return function copier (callback) {
+        cp(static, join(dest, 'shared', 'static.json'), params, callback)
       }
-    })
+    }), _done)
   }
   else callback()
 }