+++ src/shared/copy.js
@@ -1,5 +1,5 @@
 let cp = require('cpr')
-let { existsSync: exists, mkdirSync: mkdir } = require('fs')
+let { existsSync, mkdirSync } = require('fs')
 let { dirname } = require('path')
 let { sync: rm } = require('rimraf')
 let { sync: symlinkOrCopy } = require('symlink-or-copy')
@@ -9,10 +9,10 @@ module.exports = function copy (source, destination, params, callback) {
   try {
     // Toggling between symlink enabled/disabled can create weird side effects
     // Always try to remove the destination if it exists
-    if (exists(destination)) {
+    if (existsSync(destination)) {
       rm(destination)
     }
-    mkdir(dirname(destination), { recursive: true })
+    mkdirSync(dirname(destination), { recursive: true })
     if (symlink) {
       symlinkOrCopy(source, destination)
       callback()