+++ src/shared/index.js
@@ -1,18 +1,20 @@
-let fs = require('fs')
-let path = require('path')
+let { existsSync } = require('fs')
+let { join } = require('path')
 let series = require('run-series')
 let stripAnsi = require('strip-ansi')
+let getPaths = require('./get-paths')
 let copyShared = require('./copy-shared')
 let copyViews = require('./copy-views')
 let copyStaticJSON = require('./copy-static-json')
 let copyArc = require('./copy-arc')
 let { updater } = require('@architect/utils')
+let inv = require('@architect/inventory')
 
 module.exports = function shared (params = {}, callback) {
-  let { quiet, update } = params
-  let shared = path.join(process.cwd(), 'src', 'shared')
-  let views = path.join(process.cwd(), 'src', 'views')
-  let hasShared = fs.existsSync(shared) || fs.existsSync(views)
+  let { quiet, update, inventory } = params
+  let shared = join(process.cwd(), 'src', 'shared')
+  let views = join(process.cwd(), 'src', 'views')
+  let hasShared = existsSync(shared) || existsSync(views)
 
   let start
   if (!update) {
@@ -25,16 +27,32 @@ module.exports = function shared (params = {}, callback) {
 
   series([
     function (callback) {
-      copyShared(params, callback)
+      if (!inventory) {
+        inv({}, function (err, result) {
+          if (err) callback(err)
+          else {
+            inventory = params.inventory = result
+            callback()
+          }
+        })
+      }
+      else callback()
     },
     function (callback) {
-      copyViews(params, callback)
+      let paths = getPaths(inventory, 'shared')
+      copyShared(params, paths, callback)
     },
     function (callback) {
-      copyStaticJSON(params, callback)
+      let paths = getPaths(inventory, 'views')
+      copyViews(params, paths, callback)
     },
     function (callback) {
-      copyArc(params, callback)
+      let paths = getPaths(inventory)
+      copyStaticJSON(params, paths, callback)
+    },
+    function (callback) {
+      let paths = getPaths(inventory)
+      copyArc(params, paths, callback)
     },
   ], function done (err, result) {
     if (err) callback(err)